refactor: hub, routing 도메인 내 패키지 구조 정리 (#1)
