--- src/main/java/org/pokeherb/hubservice/domain/hub/CheckAccessHub.java
@@ -1,7 +1,0 @@
-package org.pokeherb.hubservice.domain.hub;
-
-import java.util.UUID;
-
-public interface CheckAccessHub {
-    void checkAccess();
-}
