+++ src/main/java/org/pokeherb/hubservice/domain/hub/service/CheckAccessHub.java
@@ -1,0 +1,5 @@
+package org.pokeherb.hubservice.domain.hub.service;
+
+public interface CheckAccessHub {
+    void checkAccess();
+}
