+++ src/main/java/org/pokeherb/hubservice/domain/routing/entity/FinalRouting.java
@@ -1,8 +1,11 @@
-package org.pokeherb.hubservice.domain.routing;
+package org.pokeherb.hubservice.domain.routing.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.pokeherb.hubservice.domain.hub.CheckAccessHub;
+import org.pokeherb.hubservice.domain.hub.service.CheckAccessHub;
+import org.pokeherb.hubservice.domain.routing.service.FinalRouteFactory;
+import org.pokeherb.hubservice.domain.routing.value.FinalTravelInfo;
+import org.pokeherb.hubservice.domain.routing.service.TravelInfoCalculator;
 import org.pokeherb.hubservice.global.domain.Auditable;
 
 import java.util.List;
