+++ src/main/java/org/pokeherb/hubservice/domain/routing/service/TravelInfoCalculator.java
@@ -1,4 +1,4 @@
-package org.pokeherb.hubservice.domain.routing;
+package org.pokeherb.hubservice.domain.routing.service;
 
 import java.util.List;
 import java.util.UUID;
